+++ packages/imperative/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Removed large dependency `pacote` for fetching package metadata when installing plug-ins. [#2668](https://github.com/zowe/zowe-cli/pull/2668)
+
 ## `5.27.13`
 
 - BugFix: Fixed truncation that could occur when the `AbstractRestClient` request method downloads gzip-encoded data. [#2661](https://github.com/zowe/zowe-cli/pull/2661)